app/requirements.txt
aiogram==3.12.0
aiohttp==3.10.5
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
